+++ src_bonus/render/intersections_cylinder.c
@@ -88,8 +88,8 @@ int	get_cylinder_inter(const t_object *object,
 	d.discriminant = (d.b * d.b) - (4 * d.a * d.c);
 	if (d.discriminant >= 0)
 	{
-		d.t0 = (-d.b - sqrtf(d.discriminant)) / (2.0 * d.a);
-		d.t1 = (-d.b + sqrtf(d.discriminant)) / (2.0 * d.a);
+		d.t0 = (-d.b - sqrt(d.discriminant)) / (2.0 * d.a);
+		d.t1 = (-d.b + sqrt(d.discriminant)) / (2.0 * d.a);
 		hits_added += store_cylinder_inter(&d, object, ray, list);
 	}
 	hits_added += intersect_caps(object, ray, list);
